execute action chains on ssh-push minions

+++ susemanager-utils/susemanager-sls/src/modules/mgractionchains.py
@@ -52,7 +52,7 @@ def _get_ac_storage_filenamepath():
 
     return os.path.join(minion_d_dir, '_mgractionchains.conf')
 
-def _read_next_ac_chunk():
+def _read_next_ac_chunk(clear=True):
     '''
     Read and remove the content of '_mgractionchains.conf' file. Return the parsed YAML.
     '''
@@ -63,19 +63,31 @@ def _read_next_ac_chunk():
     try:
         with salt.utils.fopen(f_storage_filename, "r") as f_storage:
             ret = yaml.load(f_storage.read())
-        os.remove(f_storage_filename)
+        if clear:
+            os.remove(f_storage_filename)
         return ret
     except (IOError, yaml.scanner.ScannerError) as exc:
         err_str = "Error processing YAML from '{0}': {1}".format(f_storage_filename, exc)
         log.error(err_str)
         raise CommandExecutionError(err_str)
 
+def _add_boot_time(next_chunk, prefix):
+    '''
+    Add the current boot time to the next_chunk dict
+    '''
+    uptime = __salt__["status.uptime"]()
+    next_chunk["{0}_boot_time".format(prefix)] = uptime["since_iso"]
+
 def _persist_next_ac_chunk(next_chunk):
     '''
     Persist next_chunk to execute as YAML in '_mgractionchains.conf'
     '''
+    _add_boot_time(next_chunk, "persist")
     f_storage_filename = _get_ac_storage_filenamepath()
     try:
+        f_storage_dir = os.path.dirname(f_storage_filename);
+        if not os.path.exists(f_storage_dir):
+            os.makedirs(f_storage_dir)
         with salt.utils.fopen(f_storage_filename, "w") as f_storage:
             f_storage.write(yaml.dump(next_chunk))
     except (IOError, yaml.scanner.ScannerError) as exc:
@@ -98,16 +110,19 @@ def start(actionchain_id):
     '''
     if os.path.isfile(_get_ac_storage_filenamepath()):
         msg = "Action Chain '{0}' cannot be started. There is already another " \
-              "Action Chain being executed. Please check on '{1}'".format(
+              "Action Chain being executed. Please check file '{1}'".format(
                 actionchain_id, _get_ac_storage_filenamepath())
         log.error(msg)
         raise CommandExecutionError(msg)
     target_sls = _calculate_sls(actionchain_id, __grains__['machine_id'], 1)
     log.debug("Starting execution of SUSE Manager Action Chains ID "
               "'{0}' -> Target SLS: {1}".format(actionchain_id, target_sls))
-    return __salt__['state.sls'](target_sls, queue=True)
+    ret = __salt__['state.sls'](target_sls, queue=True)
+    if isinstance(ret, list):
+        raise CommandExecutionError(ret)
+    return ret
 
-def next(actionchain_id, chunk):
+def next(actionchain_id, chunk, next_action_id=None, ssh_extra_filerefs=None):
     '''
     Persist the next Action Chain chunk to be executed by the 'resume' method.
 
@@ -123,7 +138,22 @@ def next(actionchain_id, chunk):
     yaml_dict = {
         'next_chunk': _calculate_sls(actionchain_id, __grains__['machine_id'], chunk)
     }
+    if next_action_id:
+        yaml_dict['next_action_id'] = next_action_id
+    if ssh_extra_filerefs:
+        yaml_dict['ssh_extra_filerefs'] = ssh_extra_filerefs
     _persist_next_ac_chunk(yaml_dict)
+
+def get_pending_resume():
+    '''
+    Get information about any pending action chain chunk execution.
+    '''
+    next_chunk = _read_next_ac_chunk(False)
+    if next_chunk:
+        _add_boot_time(next_chunk, "current")
+    return next_chunk or {}
+
+
 
 def resume():
     '''
@@ -150,4 +180,6 @@ def clean():
     Clean execution of an Action Chain by removing '_mgractionchains.conf'.
     '''
     _read_next_ac_chunk()
-    return True
+    return {"success": True}
+
+
